Components: Cleanup and renaming
renamed `Filereplicas` to `FileReplicas` with capital `R`
